requirements.txt
bcrypt==4.0.1
blinker==1.7.0
boto3==1.29.4
botocore==1.32.4
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
paho-mqtt==1.6.1
pip==23.3.1
pymongo==4.6.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wheel==0.41.3
zipp==3.17.0
